# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Sample API
  description: Demo REST API generated offline from Spring MVC annotations
  contact:
    name: API Team
    email: api@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: https://api.example.com/sample-api/
  description: Production
- url: https://staging.example.com/sample-api/
  description: Staging
- url: http://localhost:8080/sample-api/
  description: Local development
security:
- bearerAuth: []
...
paths:
...
  /api/v1/notifications:
    post:
      tags:
      - notifications
      summary: Send notification
      description: Dispatches a notification to the given recipient via the requested
        channel.
      operationId: sendNotification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendNotificationRequest'
        required: true
      responses:
        "202":
          description: Notification accepted for delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDto'
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDto'
        "503":
          description: No provider available for the requested channel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDto'
  /api/v1/notifications/{id}:
    get:
      tags:
      - notifications
      summary: Get notification
      description: Returns the current status and metadata of a single notification.
      operationId: getNotification
      parameters:
      - name: id
        in: path
        description: Notification ID
        required: true
        schema:
          type: integer
          format: int64
        example: "1"
      responses:
        "200":
          description: Notification found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDto'
        "404":
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDto'
    delete:
      tags:
      - notifications
      summary: Cancel notification
      description: Cancels a pending notification. Has no effect if already delivered.
      operationId: cancelNotification
      parameters:
      - name: id
        in: path
        description: Notification ID to cancel
        required: true
        schema:
          type: integer
          format: int64
        example: "1"
      responses:
        "204":
          description: Notification cancelled
        "404":
          description: Notification not found
        "409":
          description: Notification already delivered or failed
...
  /api/v1/agents/{id}/group:
    get:
      tags:
      - Generic REST API
      - Agents
      summary: Get Agent Group by Agent ID
      description: Returns the agent group associated with the given agent ID
      operationId: getAgentGroupByAgentId
      parameters:
      - name: id
        in: path
        description: Logical agent ID
        required: true
        schema:
          type: string
        example: agent-123
      responses:
        "200":
          description: Agent group found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentGroupDto'
        "404":
          description: Agent or group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentGroupDto'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentGroupDto'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentGroupDto'
...
  /api/v1/products/{id}:
    get:
      tags:
      - products
      summary: Get by ID
      description: Returns a single resource identified by its primary key.
      operationId: getById_1
      parameters:
      - name: id
        in: path
        description: Primary key of the resource
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Resource found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
    delete:
      tags:
      - products
      summary: Delete by ID
      description: Permanently removes a resource.
      operationId: deleteById
      parameters:
      - name: id
        in: path
        description: Primary key of the resource to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Resource deleted
        "404":
          description: Resource not found
...
components:
  schemas:
...
    NotificationDto:
      required:
      - recipient
      - type
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          format: int64
          readOnly: true
          example: 1
        type:
          type: string
          description: Delivery channel
          example: EMAIL
          enum:
          - EMAIL
          - SMS
          - PUSH
        recipient:
          type: string
          description: "Recipient address (email, phone number, or device token)"
          example: user@example.com
        subject:
          type: string
          description: Message subject or title
          example: Your order has shipped
        status:
          type: string
          description: Current delivery status
          example: SENT
          enum:
          - PENDING
          - SENT
          - FAILED
          - CANCELLED
        createdAt:
          type: string
          description: When the notification was created
          format: date-time
          readOnly: true
        sentAt:
          type: string
          description: When the notification was successfully delivered; null if not
            yet sent
          format: date-time
          readOnly: true
      description: Notification data transfer object
    SendNotificationRequest:
      required:
      - body
      - recipient
      - subject
      - type
      type: object
      properties:
        type:
          type: string
          description: Delivery channel
          example: EMAIL
          enum:
          - EMAIL
          - SMS
          - PUSH
        recipient:
          type: string
          description: "Recipient address (email, phone number, or device token)"
          example: user@example.com
        subject:
          type: string
          description: Message subject or title
          example: Your order has shipped
        body:
          type: string
          description: Full message body
          example: "Your order #42 is on its way!"
        priority:
          type: string
          description: Delivery priority
          example: NORMAL
          default: NORMAL
          enum:
          - LOW
          - NORMAL
          - HIGH
      description: Payload for dispatching a notification
...
    AgentGroupDto:
      required:
      - name
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the agent group
          readOnly: true
          example: group-001
        name:
          type: string
          description: Display name of the agent group
          example: Tier-1 Support
      description: Agent group data transfer object
...
    ProductDto:
      required:
      - name
      - price
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          description: Product name
          example: Wireless Keyboard
        description:
          type: string
          description: Detailed description
          example: Compact bluetooth keyboard
        price:
          type: number
          description: Unit price in EUR
          example: 49.99
        category:
          type: string
          description: Product category
          example: ELECTRONICS
          enum:
          - ELECTRONICS
          - CLOTHING
          - FOOD
          - BOOKS
        stock:
          type: integer
          description: Available stock quantity
          format: int32
          example: 150
      description: Product data transfer object
  securitySchemes:
    bearerAuth:
      type: http
      description: "JWT token obtained from the auth service (Authorization: Bearer\
        \ <token>)"
      scheme: bearer
      bearerFormat: JWT